Polish code for navigation bar
