+++ src/constants/Navigation/NavBarMenuItems.js
@@ -1,0 +1,28 @@
+const menuItems = [
+  {
+    pageTitle: "Announcement",
+    pageURL: "/announcement",
+  },
+  {
+    pageTitle: "Challenges",
+    pageURL: "/challenges",
+  },
+  {
+    pageTitle: "Rules",
+    pageURL: "/rules",
+  },
+  {
+    pageTitle: "Submission",
+    pageURL: "/submission",
+  },
+  {
+    pageTitle: "FAQ",
+    pageURL: "/faq",
+  },
+  {
+    pageTitle: "Contact",
+    pageURL: "/contact",
+  },
+];
+
+export default menuItems;
